feat: update-submodules GH action
Automatically update submodules (e.g. schemastore) if they have any
changes.

The job runs on a schedule once a month (on the Tue of the second week
of every month, at 3:42AM) and it can also be triggered manually.

The job is quite simple because all of the "magic" happens in
peter-evans/create-pull-request: this action creates and updates a PR if
there are changes in the repo with respect to the PR branch, and it also
closes the PR and deletes the PR branch if if those changes are merged
to the base branch (main) in the repo.

Fixes: #65